refactor: remove `~View` components
* use `FloatingLayer` directly as the component needs
