+++ apps/calendar/src/factory/day.tsx
@@ -1,6 +1,6 @@
 import { h } from 'preact';
 
-import DayView from '@src/components/view/dayView';
+import DayView from '@src/components/view/day';
 import CalendarControl from '@src/factory/calendarControl';
 
 export default class Day extends CalendarControl {
