+++ apps/calendar/src/factory/month.tsx
@@ -4,7 +4,7 @@
  */
 import { h } from 'preact';
 
-import MonthView from '@src/components/view/monthView';
+import MonthView from '@src/components/view/month';
 import CalendarControl from '@src/factory/calendarControl';
 import { Option } from '@src/model';
