+++ apps/calendar/stories/monthView.stories.tsx
@@ -1,6 +1,7 @@
 import { h } from 'preact';
 
-import MonthView from '@src/components/view/monthView';
+import FloatingLayer from '@src/components/floatingLayer';
+import MonthView from '@src/components/view/month';
 import { EventModelData } from '@src/model';
 import EventModel from '@src/model/eventModel';
 import TZDate from '@src/time/date';
@@ -50,6 +51,7 @@ function createMonthEvents() {
 const Template: Story = (args) => (
   <ProviderWrapper options={args.options} events={args.events}>
     <MonthView />
+    <FloatingLayer />
   </ProviderWrapper>
 );
