+++ apps/calendar/stories/weekView.stories.tsx
@@ -1,6 +1,7 @@
 import { h } from 'preact';
 
-import WeekView from '@src/components/view/weekView';
+import FloatingLayer from '@src/components/floatingLayer';
+import WeekView from '@src/components/view/week';
 import { EventModelData } from '@src/model';
 import EventModel from '@src/model/eventModel';
 import TZDate from '@src/time/date';
@@ -50,6 +51,7 @@ function createWeekEvents() {
 const Template: Story = (args) => (
   <ProviderWrapper options={args.options} events={args.events}>
     <WeekView />
+    <FloatingLayer />
   </ProviderWrapper>
 );
